Reverted to revision that didn't use BigDecimals as it didnt fix the gui shaking problem on linux

+++ src/cx/it/hyperbadger/spacezombies/explorer/SolarSystem.java
@@ -1,6 +1,5 @@
 package cx.it.hyperbadger.spacezombies.explorer;
 
-import java.math.BigDecimal;
 import java.util.ArrayList;
 
 
@@ -53,7 +52,7 @@ public class SolarSystem implements Moveable{
 	public String getName(){
 		return name;
 	}
-	public void setScale(BigDecimal scale){
+	public void setScale(double scale){
 		for(Mass m: masses){
 			m.setScale(scale);
 		}
